refactoring

+++ Repositories/RoleRepository.cs
@@ -3,6 +3,7 @@ using RoleService.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace RoleService.Repositories
 {
@@ -15,28 +16,30 @@ namespace RoleService.Repositories
             _context = context;
         }
 
-        public IEnumerable<Role> GetAll()
+        public async Task<IEnumerable<Role>> GetAll() => _context.Roles.ToList();
+
+        public async Task<Role> GetById(Guid id) => await _context.Roles.FindAsync(id);
+
+        public async Task<Role> Create(Role entity)
         {
-            return _context.Roles.ToList();
+            Role role = await _context.Roles.AddAsync(entity);
+            await _context.SaveChangesAsync();
+
+            return role;
         }
 
-        public Role GetById(Guid id)
-        {
-            return _context.Roles.Find(id);
-        }
-
-        public void Create(Role entity)
-        {
-            _context.Roles.Add(entity);
-            _context.SaveChanges();
-        }
-
-        public void Remove(Guid id)
+        public async Task<bool> Remove(Guid id)
         {
             Role role = _context.Roles.Find(id);
+            if (role == null)
+            {
+                return false;
+            }
 
             _context.Roles.Remove(role);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
+
+            return true;
         }
     }
 }
